+++ packages/exporter/index.ts
@@ -4,6 +4,7 @@ import { login } from "./login";
 
 async function main(){
 	const cookies = await login();
+	console.log("Cookies from login: ",cookies);
 	await exportData(cookies);
 }
