+++ packages/exporter/login.ts
@@ -5,36 +5,9 @@ let csrf : string;
 let requestCookies : string;
 let mobileNumber: string = "8335852033";
 
-// Global cookie jar using a Set
-const cookieJar : Set<String> = new Set();
-
-async function buildCookieHeader(response : Response, cookieNameToRemove : string | null = null) {
-	const setCookies : string[] = response.headers.getSetCookie();
-
-	if (!setCookies || setCookies.length === 0) {
-		return '';
-	}
-
-	setCookies.forEach((cookie) => {
-		const cookieValue = cookie.split(';')[0];
-		cookieJar.add(cookieValue);
-	});
-
-	if (cookieNameToRemove) {
-		cookieJar.forEach(cookie => {
-			if (cookie.startsWith(`${cookieNameToRemove}=`)) {
-				cookieJar.delete(cookie);
-			}
-		});
-	}
-	return Array.from(cookieJar).join('; ');
-}
-
-
-
 async function visitSwiggy() {
 	const response = await sw.hitURL(sw.SWIGGY_BASE_URL, sw.constOpts);
-	requestCookies = await buildCookieHeader(response);
+	requestCookies = await sw.buildCookieHeader(response);
 	const res = await response.text();
 	let regex = /window\._csrfToken = "[^"]*"/i;
 	let matches = res.match(regex)
@@ -65,7 +38,7 @@ async function generateOTP() {
 		throw new Error("Mobile number is not set in env");
 	}
 	console.log("Hitting otp gen url ...");
-	await sw.hitURL(sw.SWIGGY_GENERATE_OTP_URL, options);
+	const val = await sw.hitURL(sw.SWIGGY_GENERATE_OTP_URL, options);
 }
 
 async function performLogin() {
@@ -89,8 +62,9 @@ async function performLogin() {
 	const response = await sw.hitURL(sw.SWIGGY_LOGIN_URL, options);
 
 	// update them cookies and remove tid coz we logged in
-	requestCookies = await buildCookieHeader(response, '_guest_tid');
+	requestCookies = await sw.buildCookieHeader(response, '_guest_tid');
 	console.log(response.ok ?"Logged In" : "Login failed");
+	// console.log(response);
 	
 	return requestCookies;
 }
