+++ processed_orders.json
@@ -1,0 +1,88 @@
+[
+  {
+    "order_id": 184074106110541,
+    "order_date": "2024-08-31T11:41:46.000Z",
+    "delivery_date": "2024-08-31T12:11:54.000Z",
+    "restaurant_id": "11094",
+    "restaurant_name": "Pizza Hut",
+    "restaurant_city": "Hyderabad",
+    "restaurant_cuisine": [
+      "Pizzas"
+    ],
+    "items": [
+      {
+        "name": "Create Your Flavour Fun Combo - Box Of 2 - Non Veg Pizza - Classic Chicken Pepperoni & Onion",
+        "price": 149
+      },
+      {
+        "name": "Create Your Flavour Fun Combo - Box Of 2 - Non Veg Pizza - Chatpata Chicken Feast",
+        "price": 189
+      }
+    ],
+    "coupon_applied": "FLAT125",
+    "order_status": "Delivered",
+    "is_long_distance": false,
+    "on_time": true,
+    "rain_mode": "0",
+    "is_veg": false,
+    "is_gourmet": false,
+    "rating_meta": {
+      "restaurant_rating": 0,
+      "delivery_rating": 0
+    },
+    "item_total": "338.00",
+    "order_packing_charges": "25.00",
+    "partial_platform_fees": "6.00",
+    "delivery_charges": "FREE",
+    "discount_applied": "125.00",
+    "total_taxes": "14.54",
+    "order_total_string": "259.00",
+    "paymentMethod": "UPIIntent",
+    "paymentGateway": "googlepay"
+  },
+  {
+    "order_id": 183891849652313,
+    "order_date": "2024-08-29T09:04:09.000Z",
+    "delivery_date": "2024-08-29T09:26:52.000Z",
+    "restaurant_id": "854584",
+    "restaurant_name": "California Burrito",
+    "restaurant_city": "Hyderabad",
+    "restaurant_cuisine": [
+      "Mexican",
+      "American",
+      "Salads",
+      "Continental",
+      "Keto",
+      "Healthy Food",
+      "Beverages",
+      "Snacks",
+      "Desserts",
+      "Fast Food"
+    ],
+    "items": [
+      {
+        "name": "Grilled Barbeque Chicken Rice Bowl (Regular)",
+        "base_price": "309"
+      }
+    ],
+    "coupon_applied": "",
+    "order_status": "Delivered",
+    "is_long_distance": false,
+    "on_time": false,
+    "rain_mode": "0",
+    "is_veg": false,
+    "is_gourmet": false,
+    "rating_meta": {
+      "restaurant_rating": 0,
+      "delivery_rating": 0
+    },
+    "item_total": "309.00",
+    "order_packing_charges": "15.00",
+    "partial_platform_fees": "6.00",
+    "delivery_charges": "FREE",
+    "total_taxes": "17.28",
+    "order_total_string": "347.00",
+    "paymentMethod": "UPIIntent",
+    "paymentGateway": "googlepay"
+  }
+]
